requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Click==7.0
discord.py==1.6.0
dnspython==1.15.0
dnspython3==1.15.0
idna==2.9
idna-ssl==1.1.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mcstatus==2.3.0
multidict==4.7.5
pkg-resources==0.0.0
python-dotenv==0.11.0
six==1.14.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
